backend/requirements.txt
fastapi==0.115.0
google-genai>=1.0.0
uvicorn[standard]==0.30.6
motor==3.6.0
anthropic==0.34.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.8.2
python-multipart==0.0.9
ffmpeg-python==0.2.0
transformers>=4.45.0
torch>=2.6.0
torchaudio>=2.6.0
soundfile==0.12.1
accelerate>=0.34.2
